% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccBehav.R
\name{AccBehav}
\alias{AccBehav}
\title{Classifying behavior from activity count data}
\usage{
AccBehav(ActivityCount)
}
\arguments{
\item{ActivityCount}{A number or vector of numbers representing the 5-minute activity count data from a Telonics GPS collar.}
}
\value{
The most likely behavior for the activity count value.
}
\description{
Classify 5-minute segments of activity count data from Telonics GPS collars  
worn by mule deer into three behavioral states (Foraging, stationary, and 
Travelling). Refer to Riginos et al. (2013) Mule Deer (Odocoileus hemionus) 
Movements and Habitat Use Patterns in Relation to Roadways in Northwest 
Wyoming for details of data collection. Prediction accuracy of the function 
AccBehav() was 74 percent (95 percent C.I. 66-81%) and the area under the receiver 
operating curve (AUC) was 0.92 (95 percent C.I. 0.88-0.95)
}
\examples{
# Create toy dataset
x <- round(rlnorm(50, 3, 1))
                 
# Classify the data
AccBehav(x)

}
