% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccelerometerBehavior.R
\name{AccelerometerBehavior}
\alias{AccelerometerBehavior}
\title{Classifying behavior
Classifying behavior}
\usage{
AccelerometerBehavior(species, x, y, z, data)
}
\arguments{
\item{species}{The species ("moose", "bighorn sheep", "mule deer", or "ungulate").}

\item{x}{The column name of the X acceleration data.}

\item{y}{The column name of the Y acceleration data.}

\item{z}{The column name of the Z acceleration data.}

\item{data}{The name of the dataframe object with your data.}
}
\value{
Your dataframe object with an added column named "behavior" with a the classified behavior for each 5-minute segment.
}
\description{
Classify 5-minute segments of X, Y, Z, accelerometer data obtained from GPS 
collars from moose, bighorn sheep, mule deer, and similar ungulates into three behavioral states.
}
\examples{
# Create toy dataset
df <- data.frame(ActivityX = round(runif(10, 0, 255), 0),
                 ActivityY = round(runif(10, 0, 255), 0),
                 ActivityZ = round(runif(10, 0, 255), 0))
                 
# Classify the data
AccelerometerBehavior(species = "moose",
                      x = "ActivityX",
                      y = "ActivityY",
                      z = "ActivityZ",
                      data = df)
}
