% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstDiff2Act.R
\name{FirstDiff2Act}
\alias{FirstDiff2Act}
\title{Convert first difference of acceleration in to Vectronics activity data}
\usage{
FirstDiff2Act(x, unit)
}
\arguments{
\item{x}{A number or vector of numbers representing a 5-minute average of the first difference of acceleration from a GPS collar}

\item{unit}{What units the input data is in. Either "g" or "m/s^2".}
}
\description{
Convert the mean of absolute values of the first difference of acceleration 
in to  Vectronics activity data for use with the AccelerometerBehavior() 
function.
}
\examples{
# Create toy dataset
x <- round(rlnorm(50, 3, 1))
                 
# Convert the data to activity data
FirstDiff2Act(x, unit = "g")

}
